+++ ghmonitor/backend.py
@@ -1,3 +1,5 @@
+"""All classes inherited from the Backend can be used for sending notifications."""
+
 import logging
 import json
 
@@ -10,6 +12,8 @@ logger = logging.getLogger('Monitor')
 
 class Backend(ABC):
     """
+    Backend for sending notifications.
+
     An abstract class, that defines interface for any backend implementation. The backend is
     used for sending notifications regarding new events in the repositories, that we monitor.
     """
@@ -17,19 +21,23 @@ class Backend(ABC):
     @abstractmethod
     def notify(self, notification_string):
         # type: (str) -> None
+        """Send notification."""
         pass
 
 
 class LoggerBackend(Backend):
     """
-    Simple backend, that will use the logger for sending notifications. Useful mainly for
-    local development and testing.
+    Simple backend, that will use the logger for sending notifications.
+
+    Useful mainly for local development and testing.
     """
 
     def __init__(self):
+        """Create this backend."""
         logger.info('Using logger backend')
 
     def notify(self, notification_string):
+        """See parent class."""
         logger.info(notification_string)
 
 
@@ -37,20 +45,17 @@ SELINON_FLOW_NAME = 'golangCVEPredictionsFlow'
 
 
 class SelinonBackend(Backend):
-    """
-    Production backend that is connected to Selinon(Celery(SQS@AWS))
-    """
-
-    def __init__(self):
-        pass
+    """Production backend that is connected to Selinon(Celery(SQS@AWS))."""
 
     def notify(self, notification_string):
+        """See parent class."""
         init_celery(result_backend=False)
         run_flow(SELINON_FLOW_NAME, notification_string)
 
 
 def create_pr_notification(package, repository, id):
     # type: (str, str, int) -> str
+    """Create notification about new PR encoded into JSON string."""
     notification_dict = {
         "repository": repository,
         "package": package,
@@ -62,6 +67,7 @@ def create_pr_notification(package, repository, id):
 
 def create_issue_notification(package, repository, id):
     # type: (str, str, int) -> str
+    """Create notification about new issue encoded into JSON string."""
     notification_dict = {
         "repository": repository,
         "package": package,
@@ -73,6 +79,7 @@ def create_issue_notification(package, repository, id):
 
 def create_push_notification(package, repository):
     # type: (str, str) -> str
+    """Create notification about new push event encoded into JSON string."""
     notification_dict = {
         "repository": repository,
         "package": package,
