+++ ghmonitor/gopkg/__init__.py
@@ -1,1 +1,1 @@
-# ...
+"""Utilities specific to Go language ecosystem."""
