+++ ghmonitor/gopkg/translate.py
@@ -1,3 +1,5 @@
+"""Utility functions for translation of Go packages into Github repositories."""
+
 import re
 from requests_html import HTMLSession
 
@@ -10,6 +12,8 @@ from typing import Union
 def get_repo_from_random_urn(urn):
     # type: (str) -> Union[str, None]
     """
+    Translate Go package urn into Github repo.
+
     In case the repository is not directly from GitHub, it is possible that there is a way to
     translate the urn into Github repository using some magic from "go get". So just try it.
     """
@@ -44,6 +48,7 @@ def translate(pkg):
     # type: (str) -> Union[str, None]
     """
     Take whatever `go get` takes and return string with organization/user and repository name.
+
     It returns none for anything that is not available on Github.
     """
     # Test if the pkg is directly from github
@@ -57,25 +62,3 @@ def translate(pkg):
         return repo
 
     return None
-
-
-def test_get_repo_urn():
-    assert 'kubernetes/metrics' == get_repo_from_random_urn('k8s.io/metrics')
-    assert get_repo_from_random_urn('seznam.cz') is None
-
-
-def test_github_re():
-    assert GITHUB_REPO_RE.match('bitbucket.com/user/repo') is None
-    m = GITHUB_REPO_RE.match('github.com/user/project')
-    assert m.group('user') == 'user'
-    assert m.group('repo') == 'project'
-    m = GITHUB_REPO_RE.match('github.com/user/project/folder')
-    assert m.group('user') == 'user'
-    assert m.group('repo') == 'project'
-
-
-def test_translate():
-    assert 'kubernetes/metrics' == translate('k8s.io/metrics')
-    assert 'user/project' == translate('github.com/user/project')
-    assert translate('launchpad.net/project') is None
-
