+++ ghmonitor/models.py
@@ -1,3 +1,5 @@
+"""Classes representing Github events and associated stuff."""
+
 import logging
 
 from enum import Enum
@@ -6,16 +8,14 @@ logger = logging.getLogger('Monitor')
 
 
 class UnknownEvent(Exception):
-    """
-    Exception for unknown event strings.
-    """
+    """Exception for unknown event strings."""
+
     pass
 
 
 class EventType(Enum):
-    """
-    Represent event types that we are interested in like new pull requests, issues, or comments.
-    """
+    """Represent event types that we are interested in (new pull requests, issues, or comments)."""
+
     PUSH = 1
     PULL_REQUEST = 2
     ISSUE = 3
@@ -23,6 +23,7 @@ class EventType(Enum):
     @staticmethod
     def from_str(input_string):
         # type: (str) -> EventType
+        """Create an event type from the string or raise the UnknownEvent exception."""
         mapping = {
             "PushEvent": EventType.PUSH,
             "PullRequestReviewCommentEvent": EventType.PULL_REQUEST,
@@ -38,29 +39,33 @@ class EventType(Enum):
 class Event:
     """
     Event structure that contains only information relevant for our use case.
+
     Also performs type check.
     """
+
     def __init__(self):
+        """Create new event."""
         self.id = None
         self.type = None
         self.repo = None
 
     def __str__(self):
+        """Return pretty formatted string describing the Event object."""
         return "<Event id='{}', type='{}', repo='{}'>".format(self.id, self.type, self.repo)
 
     def __eq__(self, other):
+        """Needed for operations in a set."""
         return (self.id, self.type, self.repo) == (other.id, other.type, other.repo)
 
     def __hash__(self):
-        """
-        https://docs.python.org/3/reference/datamodel.html#object.__hash__
-        """
+        """See https://docs.python.org/3/reference/datamodel.html#object.__hash__ for more info."""
         return hash((self.id, self.type, self.repo))
 
     @staticmethod
     def from_dict(input_dict):
         """
-        Filter the input dictionary and type check its members
+        Filter the input dictionary and type check its members.
+
         :param input_dict: JSON as a Python dictionary
         :return: Event object or None in case of unknown event type or other failure
         """
@@ -79,35 +84,3 @@ class Event:
         except KeyError:
             logger.error('Input dictionary does not contain required keys')
             return None
-
-
-def test_event_types_from_str():
-    assert EventType.PUSH == EventType.from_str("PushEvent")
-    assert EventType.ISSUE == EventType.from_str("IssuesEvent")
-    assert EventType.PULL_REQUEST == EventType.from_str("PullRequestEvent")
-    try:
-        EventType.from_str("foobar")
-    except Exception as e:
-        assert isinstance(e, UnknownEvent)
-
-
-def test_event_parser_returns_event():
-    print("Testing data parser")
-    a = {"id": "222", "type": "PushEvent", "repo": {"name": "a"}}
-    b = {"id": "222", "tpe": "PushEvent", "repo": {"name": "a"}}
-    assert isinstance(Event.from_dict(a), Event)
-    assert Event.from_dict(b) is None
-
-
-def test_events_comparison():
-    e0 = Event()
-    e1 = Event()
-    assert e0 == e1
-    e0.repo = 'a'
-    assert not(e0 == e1)
-    e1.repo = 'a'
-    e0.id = 1
-    e1.id = 1
-    e0.type = EventType.PUSH
-    e1.type = EventType.PUSH
-    assert e0 == e1
